+++ internal/config/file.go
@@ -113,6 +113,13 @@ type GeoIPConfig struct {
 
 // PrometheusConfig extra prometheus settings.
 type PrometheusConfig struct {
+	// Disable the collector that exports metrics about the current Go process and runtime
+	DisableGoCollector bool `json:"disable_go_collector"`
+
+	// Disables the collector that exports metrics about the current state of the process, including
+	// CPU, memory, and file descriptor usage, as well as the process startup time.
+	DisableProcessCollector bool `json:"disable_process_collector"`
+
 	// ConstantLabels are added to every metric exposed by this exporter.
 	// WARNING: changing labels creates new time series.
 	ExtraLabels map[string]string `json:"extra_labels"`